Block PRs with "high" level slither issues (#89)
This fixes the current "high" level issue flagged by slither and
prevents new ones from getting introduced.

A later iteration could take this down to "medium".

Includes a new slither workflow just for PRs that fails on high level
issues. The existing one doesn't fail, in order to produce a security
report for the main branch of the repo.