feat: multiple allowed cash tokens (#143)
- `DualGovernorDeployer` is now deployed with an array of
`allowedCashTokens_`, which are a permanent set for the system.
- This array is converted into a truthy hashmap in the `DualGovernor`
storage in its constructor.
- The first cash token in the array of `allowedCashTokens_` is the
default for any new `DualGovernor` deployment (i.e. post-reset).
- `Registrar` no longer needs a `cashToken` in its constructor, as it is
read when needed form the `DualGovernorDeployer`.
- `setCashToken(address newCashToken_)` is now a Zero type proposal.
- `PowerToken` has a method `setNextCashToken` that only its governor
can call, which ensures the new cash token takes effect only in the next
epoch and onwards, guaranteeing the continued and consistent expected
operation of the in-progress auction.